Ported to x64 (not fully tested).
Reviewed ".props" files.
Added projects for MSVC 2012.
Added configuration for x64 builds (MSVC 2012 only).
Ported to x64 (not fully tested).
Removed ET's global exception handling (win_eh.cpp).

+++ src/game/q_shared.cpp
@@ -1258,7 +1258,7 @@ char *Q_strupr( char *s1 ) {
 
 // never goes past bounds or leaves without a terminating 0
 void Q_strcat( char *dest, int size, const char *src ) {
-	int l1;
+	intptr_t l1;
 
 	l1 = strlen( dest );
 	if ( l1 >= size ) {
