+++ src/ui_mp/ui_main.cpp
@@ -286,7 +286,26 @@ qboolean _UI_IsFullscreen( void );
 #if defined( __MACOS__ )
 #pragma export on
 #endif
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
+
+//BBi
+//int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
+intptr_t vmMain (
+    intptr_t command,
+    intptr_t arg0,
+    intptr_t arg1,
+    intptr_t arg2,
+    intptr_t arg3,
+    intptr_t arg4,
+    intptr_t arg5,
+    intptr_t arg6,
+    intptr_t arg7,
+    intptr_t arg8,
+    intptr_t arg9,
+    intptr_t arg10,
+    intptr_t arg11)
+{
+//BBi
+
 #if defined( __MACOS__ )
 #pragma export off
 #endif
